refactor: move GetImageMetadata to new pkg
