+++ cmd/manifest/init-v1.go
@@ -33,7 +33,7 @@ import (
 	oktetoLog "github.com/okteto/okteto/pkg/log"
 	"github.com/okteto/okteto/pkg/model"
 	"github.com/okteto/okteto/pkg/okteto"
-	"github.com/okteto/okteto/pkg/registry"
+	"github.com/okteto/okteto/pkg/registry/registry"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 	"k8s.io/client-go/kubernetes"
 )
@@ -107,7 +107,7 @@ func (*ManifestCommand) RunInitV1(ctx context.Context, opts *InitOpts) error {
 		return err
 	}
 
-	dev, err := linguist.GetDevDefaults(opts.Language, opts.Workdir, &registry.ImageConfig{})
+	dev, err := linguist.GetDevDefaults(opts.Language, opts.Workdir, registry.ImageMetadata{})
 	if err != nil {
 		return err
 	}
