+++ cmd/manifest/init-v2.go
@@ -37,7 +37,7 @@ import (
 	oktetoLog "github.com/okteto/okteto/pkg/log"
 	"github.com/okteto/okteto/pkg/model"
 	"github.com/okteto/okteto/pkg/okteto"
-	"github.com/okteto/okteto/pkg/registry"
+	"github.com/okteto/okteto/pkg/registry/registry"
 	"github.com/spf13/afero"
 	"github.com/spf13/cobra"
 )
@@ -357,7 +357,7 @@ func (mc *ManifestCommand) configureDevsByResources(ctx context.Context, namespa
 	return nil
 }
 
-func setFromImageConfig(dev *model.Dev, imageConfig *registry.ImageConfig) {
+func setFromImageConfig(dev *model.Dev, imageConfig registry.ImageMetadata) {
 	if len(dev.Command.Values) == 0 && len(imageConfig.CMD) > 0 {
 		dev.Command = model.Command{Values: imageConfig.CMD}
 	}
@@ -584,7 +584,7 @@ func configureAutoCreateDev(manifest *model.Manifest) error {
 		return err
 	}
 
-	dev, err := linguist.GetDevDefaults(language, wd, &registry.ImageConfig{})
+	dev, err := linguist.GetDevDefaults(language, wd, registry.ImageMetadata{})
 	if err != nil {
 		return err
 	}
